# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sinapse Service Contracts
  version: 0.1.0
  description: Contratos versionados entre backend, serviço de IA e automações n8n.
servers:
  - url: http://localhost:3001
    description: Backend Node.js
  - url: http://localhost:8000
    description: Serviço de IA local
paths:
...
  /api/v1/projects:
    post:
      summary: Cadastra um novo projeto (PBI-01.1.1 / S1-03)
      operationId: createProject
      description: Escrita permitida apenas aos perfis po e admin; dev recebe 403.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Projeto criado com sucesso com status ativo e auditoria registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          description: Perfil sem permissão de escrita
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Conflito de nome duplicado entre projetos ativos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Lista projetos com paginação e busca (S1-03)
      operationId: listProjects
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Ausente exclui arquivados; arquivado retorna somente arquivados; todos inclui todos os estados.
          schema:
            type: string
            enum: [ativo, em_andamento, concluido, arquivado, todos]
        - name: busca
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, nome_asc, nome_desc]
            default: created_at_desc
      responses:
        '200':
          description: Lista paginada de projetos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/chat/projects:
    get:
      summary: Lista projetos permitidos no chat (admin/PO ou alocação ativa do desenvolvedor)
      operationId: listChatProjects
      security: [{cookieAuth: []}]
      responses:
        '200': {description: Lista items com id e nome dos projetos}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/chat/conversations:
    get:
      summary: Lista as conversas do usuário autenticado
      operationId: listChatConversations
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Conversas do usuário, da mais recente para a mais antiga
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Cria uma conversa (opcionalmente escopada a um projeto)
      operationId: createChatConversation
      security: [{cookieAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                titulo: {type: string, maxLength: 120}
                projeto_id: {type: string, format: uuid}
      responses:
        '201':
          description: Conversa criada
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Projeto não encontrado
  /api/v1/chat/conversations/{id}/messages:
    get:
      summary: Mensagens de uma conversa do próprio usuário
      description: Conversa de outra pessoa retorna 404, sem revelar a existência.
      operationId: listChatMessages
      security: [{cookieAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Mensagens em ordem cronológica
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversa não encontrada
  /api/v1/chat/query:
    post:
      summary: Envia uma pergunta ao assistente e registra a troca
      description: >
        Projeto obrigatório em novas conversas; em conversas existentes o projeto é imutável.
        Somente admin/PO ou desenvolvedor com alocação ativa pode consultar o projeto.
        Sem evidência no acervo, responde "Informação não encontrada no acervo
        do projeto." sem inventar fontes. `origem` informa se a resposta veio
        do assistente, de busca textual (assistente indisponível) ou sem resultado.
      operationId: askChatQuestion
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pergunta]
              properties:
                pergunta: {type: string, minLength: 1, maxLength: 2000}
                conversa_id: {type: string, format: uuid}
                projeto_id: {type: string, format: uuid}
      responses:
        '200':
          description: Resposta com fontes
          content:
            application/json:
              schema:
                type: object
                required: [conversa_id, resposta, fontes, origem]
                properties:
                  conversa_id: {type: string, format: uuid}
                  resposta: {type: string}
                  fontes:
                    type: array
                    items:
                      type: object
                      properties:
                        id: {type: string}
                        titulo: {type: string}
                        tipo: {type: string}
                  origem: {type: string, enum: [assistente, busca_textual, sem_resultado]}
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversa (de outro usuário incluída) ou projeto não encontrado
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sinapse_session
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [nome, cliente]
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 255
        cliente:
          type: string
          minLength: 1
          maxLength: 255
        descricao:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, em_andamento, concluido]
          default: ativo
        data_inicio:
          type: string
          format: date-time
          nullable: true
...
    ProjectResponse:
      type: object
      required: [id, nome, cliente, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        cliente:
          type: string
        descricao:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, em_andamento, concluido, arquivado]
        data_inicio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        epicos_count:
          type: integer
        archived_at:
          type: string
          format: date-time
          nullable: true
          description: Data do arquivamento; nula em registros legados sem data conhecida.
        documentos_count:
          type: integer
    PaginatedProjectsResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
  responses:
    ValidationError:
      description: Corpo inválido ou campo obrigatório ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Autenticação ausente, sessão inválida ou sessão expirada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'